MIPS: jz4780: Jump to smp_bootstrap in KSEG0 in SMP entry
The SMP entry code executes from KSEG1, and was using a PC-region
branch instruction to jump to smp_bootstrap(). This resulted in the
kernel initialisation and the idle task on the second core being
executed from KSEG1.

This fixes a few unexplained bits of behaviour on the second core, such
as the incorrect delay loop calibration, and the apparent slowness in
the idle loop which could be seen in an ftrace function trace (and on
the CPU activity LEDs on the new Ci20 boards).

Signed-off-by: Alex Smith <alex.smith@imgtec.com>