MIPS: qi_lb60: Migrate to devicetree
Move all the platform data to devicetree.

The only bit dropped is the PWM beeper, which requires the PWM driver
to be updated. I figured it's okay to remove it here since it's really
a non-critical device, and it'll be re-introduced soon enough.

The MMC max frequency was also updated to 48 MHz since the controller
supports it.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>