IIO: Ingenic JZ47xx: Set clock divider on probe
The SADC component can run at up to 8 MHz (on JZ4725B in any case),
but is fed a 12 MHz input clock (EXT). So we divide it by two to get
6 MHz. Then we also have to set up another divider to match, to
produce a 10us clock.

If the clock dividers are left on their power-on defaults (a divider
of 1), the SADC mostly works, but will occasionally produce erroneous
readings. This led to button presses being detected out of nowhere on
the RS90 every few minutes. With this change, no ghost button presses
were logged in almost a day worth of testing.

Signed-off-by: Artur Rojek <contact@artur-rojek.eu>