rtc: jz4740: Register clock provider for the CLK32K pin
On JZ4770 and JZ4780, the CLK32K pin is configurable. By default, it is
configured as a GPIO in input mode, and its value can be read through
GPIO PD14.

With this change, clients can now request the 32 kHz clock on the CLK32K
pin, through Device Tree. This clock is simply a pass-through of the
input oscillator's clock with enable/disable operations.

This will permit the WiFi/Bluetooth chip to work on the MIPS CI20 board,
which does source one of its clocks from the CLK32K pin.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

---
v3: - Use dev_err_probe()
    - Use __clk_hw_get() to get a pointer to the parent clock, instead
      of doing it by name.
    - Add Kconfig dependency on CONFIG_COMMON_CLK
    - Register CLK32K clock if the #clock-cells device property is
      present, instead of doing it based on the compatible string